# Perf/cache string ops