* `getitem` was uncached and gets hit often.
* only compute hash once per Node. This will become important once we start caching node operations (incoming)

this branch
```
codegen         mean runtime:  140.45ms, runs:   159.52,  127.21,  160.59,  134.36,  132.79,  130.65,  128.03,  134.19,  131.52,  165.66
methodcache     mean runtime:  132.99ms, runs:   132.94,  130.45,  122.93,  124.74,  130.47,  124.58,  120.56,  120.10,  184.96,  138.16
profile         mean runtime:  568.68ms, runs:   657.05,  546.74,  538.03,  552.64,  589.43,  559.45,  548.08,  550.50,  556.92,  587.93

706.71 lazy.py
557.83 tensor.py
154.16 shapetracker.py
13.52 ops.py
47.79 symbolic.py
374.19 helpers.py
8.62 lib.py
1862.81 total
```

master


```
codegen         mean runtime:  143.37ms, runs:   165.84,  141.58,  168.71,  132.82,  137.16,  131.48,  130.21,  128.91,  130.37,  166.60
methodcache     mean runtime:  132.13ms, runs:   130.96,  134.70,  129.90,  125.40,  125.76,  124.29,  124.63,  124.62,  169.28,  131.80
profile         mean runtime:  589.58ms, runs:   671.35,  580.40,  595.32,  559.58,  596.33,  569.90,  565.23,  619.70,  579.99,  557.96

721.38 lazy.py
572.08 tensor.py
155.33 shapetracker.py
14.80 ops.py
54.84 symbolic.py
385.14 helpers.py
9.43 lib.py
1913.00 total
```